refactor(app-router): keep AST utilities canonical
plugins/ast-utils.ts is already part of the prefer-shared-utils lint rule and is the canonical home for shared AST traversal helpers. Moving it behind utils/ast changed that contract unnecessarily.

Restore plugins/ast-utils.ts as the canonical helper module, remove utils/ast.ts, and have the route client-reference scanner import the existing shared AST helpers directly.
